+++ src/Geopilot.Pipeline/Pipeline.cs
@@ -64,6 +64,11 @@ public sealed class Pipeline : IPipeline
             {
                 return ProcessingState.Running;
             }
+            else if (stepStates.Contains(StepState.Warning)
+                && stepStates.All(s => s == StepState.Success || s == StepState.Skipped || s == StepState.Warning))
+            {
+                return ProcessingState.Warning;
+            }
             else if (stepStates.All(s => s == StepState.Success || s == StepState.Skipped))
             {
                 return ProcessingState.Success;